requirements.txt
fastapi
sqlalchemy
uvicorn
pymysql
python-dotenv
email_validator
cryptography
httpx
python-multipart
python-jose
passlib[bcrypt]
pytest
pytest-mock